Add ClEventsPage component and update TypeScript dependencies
